refactor(withdraw): one home for the scanned-destination contract
Smell pass on my own diff: the amount step's token pick was a second copy of
the rule WithdrawMethodView already applies to an address-book tap — USDC where
the chain has it, else the chain's only token, which is what makes Tron deliver
USDT. Two copies of that is how a chain fact drifts.

destination.ts now owns the whole contract (the URL, the validation, the token
rule) and routes.ts goes back to route builders. Both seed sites read the same
function.
