+++ src/components/Global/QRScannerOverlay/index.tsx
@@ -20,7 +20,7 @@ import { pixKeyToQrPayUrl } from '@/utils/pix.utils'
 import { extractPaymentValue } from '@/utils/clipboard-extract.utils'
 import { recipientPayUrl, qrClaimUrl, deepLinkToNativePath } from '@/utils/native-routes'
 import { qrTelemetry, reportQrScanError } from '@/components/Global/QRScanner/utils'
-import { readWithdrawDestination, withdrawDestinationUrl } from '@/features/withdraw/routes'
+import { readWithdrawDestination, withdrawDestinationUrl } from '@/features/withdraw/destination'
 import { useChainRollout } from '@/hooks/useChainRollout'
 import { useTranslations } from 'next-intl'
 import { usePathname, useRouter, useSearchParams } from 'next/navigation'
